CDP: split session.rs into per-domain sub-modules
